refactored native specific files from http to server dir
